Upgrade keeps xp stored
Now when you upgrade your book, it keeps your previously stored experience in the upgraded one. Add getter for maxExperience as I need it for the mixin. Don't use the onCraft method anymore, replaced with the mixin (so now you can see directly in the crafing table the book with the right experience in.