Can't add enchants to xp book via anvil
XP Books are simple item, but damageable so they fit the enchantment type: breakable. This means the only enchants we can add to an xp book are mending and unbreaking, which would be stupid to add but better than we can't. As the unbreaking enchant class is stupid, need 2 mixins instead of 1 for the enchant type